pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "product-roadmap-agents"
version = "0.1.0"
description = "Local multi-agent product roadmap workflow using OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.0.0",
  "openai-agents>=0.2.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
roadmap-agents = "product_roadmap_agents.cli:main"

[tool.setuptools.packages.find]
include = ["product_roadmap_agents*"]
